refactor(cli): move the benchmark group into its own module
Cuts `gpio benchmark` out of cli/main.py into
geoparquet_io/cli/commands/benchmark.py, unchanged. Clean cut: no shared
helpers and no test patch sites to repoint.

`geoparquet_io.cli.main` carries a mypy per-module override, so one
pre-existing `arg-type` error in `benchmark suite` was suppressed by living in
main.py and surfaces the moment the code moves. The override entry follows the
code rather than the error being fixed here, keeping this a pure move.

tests/data/cli_surface.json is byte-identical.
